Formulate elliptic system in terms of fluxes and sources
Refactors the elliptic code to a formulation that resembles closely the
flux-conservative form of evolution systems. The system just computes
the fluxes and sources and no longer has to implement the derivative
operator or the normal-dot-fluxes. Instead, they use code shared
between systems.

Also generalizes the elliptic internal penalty flux to be used by any
elliptic system in first-order flux-formulation.