Use Poisson field gradient for auxiliary variable
It was a mistake to use a vector quantity for the auxiliary variable before. It should be
a covector, which the principal symbol raises to a vector.
We don't need a special tag for this anymore, since we can just use
the gradient prefix.
This change also allows adding fluxes for the Poisson equation on a
non-Euclidean background metric.