Split flux communication and DG boundary scheme
This separates the flux communication code into the actions that do the
communication and the DG boundary scheme that defines what needs
to be communicated. In addition to cleaning up the flux communication
code, this enables communicating fluxes for multiple sets of variables.
It also allows to implement other DG schemes, such as a weak scheme
or a second-order elliptic scheme, and share the flux communication
infrastructure.