Rename SimpleBoundaryData to SimpleMortarData
Renaming the class because we call "boundary data" the data on
one face of an element that can be projected to a mortar,
and "mortar data" the collection of (projected) boundary data from
both element faces that touch a mortar. By this definition, the
current class is "mortar data" and is therefore renamed
`SimpleMortarData`. In a following commit a new class is added for
"boundary data" and will be named `SimpleBoundaryData`.