Derive elliptic solutions from base classes
The classes use the registration mechanism so the executables can choose
which set of classes they want to use. Because of the extra template
parameter the implementations of the solution variables have to move to
a helper class so they can be template-instantiated in a cpp file. This
has the benefit that it encourages using the CachedTempBuffer data
structure, which helps with re-using intermediate quantities.