Add tag trait db::is_mutable_item_tag
The implementation of DataBox naturally wants to split its tags_list
into tags for mutable and immutable items.  Note that currently
is_mutable_item_tag is equivalent to is_simple_tag, but this allows
for the introduction of other kinds of mutable_item_tags in the future,
and makes the code logic for DataBox more clear.