Fix round-robin element allocation in elliptic executables
Previously the distribution of elements to CPUs was reset for each
block, so a domain with only single-element blocks were was not
parallelized at all. This wasn't really an issue because for production
runs I have been using my multigrid branch with its own strategy
of distributing elements. But it would be good to fix this on develop
until the multigrid solver is merged.