Use Options::Auto for optional boundary conditions
Since Options::Auto supports pointer conversions now (#2930)
we can use it in place of the generic `None` boundary condition
to indicate the absence of a boundary condition. In the case of the
binary domain creator an inner boundary condition needs to be
imposed only if at least of of the spheres is excised. So the user
should set the inner boundary condition to None if they don't excise
a sphere.