Optionally install missing Py dependencies in build dir
When the optional `INSTALL_PY_DEPS` CMake flag is enabled, missing
Python dependencies are installed to the CMake build directory.
This is an alternative if the user doesn't want to bother installing
the dependencies themselves, e.g., in a Python venv. Packages like
h5py, numpy and scipy can/should still be installed by the user
to make sure they use the correct HDF5, BLAS, etc., but the new
`INSTALL_PY_DEPS` flag makes it easy to install pure-Python packages.