Allow monopole and dipole in shape map
These were asserted to be zero before, but the map works just as well
if these modes are nonzero.
This allows to use the shape map to reshape a coordinate sphere to
a Kerr horizon without using an additional SphericalCompression map
just for the monopole.
If the zero-ness of the monopole and dipole are diagnostics for the
control system, then they should be checked by the control system
instead of imposing this artificial constraint.