Paralelize computing vars to interpolate to AH
This PR parallelizes the computation of the variables to be interpolated
to the apparent horizon from the evolved variables. Previously,
this calculation was done for each element in serial on the
apparent horizon finder singleton. This PR makes one main change:
instead of first passing the evolved variables to the apparent horizon
finder and then computing the variables to interpolate to the horizon,
instead this PR now first computes the variables to interpolate on
each element and then passes these to the apparent horizon finder.

The reason for this change is that this computation is a significant
fraction of the time spent by the apparent horizon finder. By moving
this computation to each element, this increases the run speed in
binary black hole simluations by reducing the amount of time spent
bottlenecked by the apparent horizon finder.

The rest of the changes in this PR follow from this main change; e.g.,
changing the signature of ah::FindApparentHorizon requires some
changes to ensure that the code builds and the tests pass.